+++ src/config.js
@@ -3,7 +3,14 @@
 *
 * Application configuration
 */
-export default {
-  allowAnonymousAuth: true,
-  appTitle: "GroundLevel - sample app",
+const allowAnonymousAuth = true
+const appTitle = "GroundLevel - sample app"
+
+// as used by the browser. THIS MUST MATCH with what's in 'functions/index.js'
+const functionsRegion = "europe-west3"
+
+export {
+  allowAnonymousAuth,
+  appTitle,
+  functionsRegion
 }