+++ vite.config.js
@@ -2,5 +2,6 @@ module.exports = {
   alias: {
     // does not work. #later
     //'/@components': path.resolve(__dirname,'.')    // aim: '@' or '/@/' to point to source root
-  }
+  },
+  minify: false,
 }